backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sse-starlette>=2.0.0
python-multipart>=0.0.6

# LLM Clients
openai>=1.12.0
anthropic>=0.18.0
httpx>=0.27.0,<0.28.0

# Async Support
aiohttp>=3.9.3

# Data Validation
pydantic>=2.6.0

# Database / ORM
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Optional: MySQL support (uncomment if needed)
# pymysql>=1.1.0

# Optional: PostgreSQL support (uncomment if needed)
# psycopg2-binary>=2.9.9

# Utils
python-dotenv>=1.0.1
rich>=13.7.0

# Authentication
PyJWT>=2.8.0
bcrypt>=4.1.0

# Web Search
ddgs>=7.0.0

# Type Hints
typing-extensions>=4.9.0
